Improve implementation of SOFT.
Now DEFUN2 directly uses DEFUN and not DEFINE.

The original use of DEFINE was motivated by its :T-PROOF option for generating a
termination theorem that could be used when instantiating the 2nd-order
function. But after the :TERMINATION-THEOREM feature was made available for
every DEFUN function in ACL2, DEFINE is no longer necessary for that purpose.