Improve implementation of DEFUN-INST.
When instantiating a plain second-order function, set :VERIFY-GUARDS for the
instance to be either the value of the keyed option (if present), or based on
whether the second-order function is guard-verified or not (i.e. "copy" the
guard verification status in this case). Avoid the use of DEFINE, generating
directly a DEFUN instead.

When instantiating a choice second-order function, do not bother passing the
options through, since they are checked to be null.

When instantiating a quantifier second-order function, generate the needed
options more explcitly.