Make no-sqrt.lisp certifiable in all ACL2 versions
By replacing realp with real/rationalp and complexp with
complex/complex-rationalp, this book becomes certifiable with standard
ACL2.  As far as I've seen, usually this is done to make
ACL2-certifiable books become ACL2(r)-certifiable rather than vice
versa, but I think it's nice to try to make ACL2(r)-related books
ACL2-certifiable if possible, too.

The final theorem says that if x^2 = 2, then x is real/rationalp and
not rationalp.  In ACL2(r) this means that x is an irrational real
number, but in standard ACL2, this means that x is both rational and
not rational, which is impossible (i.e. such an x cannot exist).  So
for the standard ACL2 case (#-non-standard-analysis), I make the
theorem a corollary which simply states that there is no x such that
x^2 = 2.

The comments in this file need to be fixed as well -- they were
written in a way that implied that ACL2(r) had become the mainstream
ACL2 distribution, whereas in fact it is (currently) maintained in
parallel with a standard version of ACL2 that does not include
non-standard analysis support.  I'll do that in a separate commit.