Revert special handling of prog2$ by directed-untranslate.
I'm reverting a recent change that I made to directed-untranslate.  It
had handled a given uterm = (prog2$ u v) that simplified to a term v'
by producing (prog2$ u v'), but u could then be invalid because it's
not in the scope of the same let-bindings as in the original term.