Installed meta-, as its normal command for every Emacs.  Tweaked some :doc and a comment.
Quoting :doc note-7-5:

  Defined `meta-,' in emacs/emacs-acl2.el to be Emacs function
  tags-loop-continue, which is how it has traditionally been defined
  by Emacs but might be defined differently in some versions of Emacs
  25 (and perhaps later). NOTE: Put (setq
  *preserve-tags-loop-continue* t) in your .emacs file before loading
  ACL2 file emacs/emacs-acl2.el, if you want to avoid redefining
  `meta-,'.  Thanks to Keshav Kini and Mihir Mehta for helpful
  discussions.

Clarified in :doc acl2-doc that completion is done in the usual way,
so if you want to enter the "?" character you need to prefix it with
control-q.  Thanks to Mihir Mehta for bringing this up.

Improved comment in source function defun-inline about why inlining
works how it does in ACL2.  Thanks to Keshav Kini for helpful
feedback.