Fixed misc/profiling.lisp to certify with github distributions of CCL (addressing GitHub Issue #744).
Github distributions of CCL do not include the contrib/ directory,
which up till now has been assumed to exist by the book,
books/misc/profiling.lisp.  My fix for this is a bit ugly, but
adequate.  If there are additional such issues in the future I might
try to make a prettier fix, perhaps switching to the use of
include-raw.  It didn't seem worth the effort to do that at this
point.

I've had correspondence with a CCL developer.  If he decides to
include contrib/ in future CCL distributions, then I'll probably
revert this change; but it seems more likely that he'll put contrib/
into its own github repository.  Attempts to use github CCL with the
misc/profiling book now result in messages pointing to that future
possibility.