[x86isa] las-to-pas does not traffic in lists of canonical addresses anymore
Instead of a list of canonical addresses, a contiguous linear memory
region is specified by <n, lin-addr>, where n is the number of bytes
to be read and lin-addr is the first address of the memory region.
Note that this does limit address translation using las-to-pas to
contiguous linear memory regions at a time.  However, there's rarely
any call to do address translation for non-contiguous linear memory
regions, so this is probably fine.

It'd be nice if las-to-pas could analogously traffic in physical
memory regions --- that is, instead of trafficking in lists of
physical addresses, <n, phy-addr> could specify a region of physical
memory that corresponds to <n, lin-addr>.  However, this is not quite
so straightforward because a contiguous linear memory region can map
to multiple non-contiguous physical memory regions, depending on the
page table configuration.