Fix bug in MSG-DOWNCASE-FIRST and MSG-UPCASE-FIRST.
Before this fix, these two functions did not properly handle messages whose
format string starts with certain tilde-directives, e.g. (MSG "~s0" "string").

Now several kinds of tilde-directives are properly handled, except for a few
that are explicitly not supported. When these unsupported tilde-directives are
encountered, a hard error is generated; support for them can be added when
needed.

The documentation of these two functions has been expanded to explain the
handling of the tilde-directives, and to point out the unsupported ones.

Several tests have been added as well.