Introduce macro to set the FAULT field.
The new macro !!FAULT-FRESH is analogous to the existing !!MS-FRESH, but it sets
the FAULT field instead of the MS field.

Since we decided that the exceptions prescribed by the Intel manuals should be
stored into the FAULT field instead of the MS field (since the latter is
specific to the formal model, e.g. to signal instructions not currently
modeled), the new macro can be used in place of the old one where needed.