Improve modeling of some x86 exceptions.
For the PUSH/POP instructions, some of the exceptions prescribed by the Intel
manuals are now put into the FAULT (which describes pending exceptions) field
instead of the MS field (which is specific to the formal model).

These exceptions start with keywords :UD, :SS, etc. that correspond to #UD, #SS,
etc. Each such keyword is followed by the code (e.g. 0 for #SS(0)), or NIL if
there is no code (e.g. for #UD).

More exceptions can be handled similarly.