Add new UBODY system utility.
This is similar to the built-in BODY, but it has a stronger guard (it only
operates on logic-mode defined functions) and always returns the unnormalized
body (hence the 'U' in the name).

(UBODY FN WRLD) can be used in place of (BODY FN NIL WRLD), avoiding the NIL
argument. UBODY is guard-verified, unlike (currently) BODY.