Improve DEF-ERROR-CHECKER utility.
This macro generates a function (and an associated macro) that includes a formal
called DESCRIPTION. Previously, this symbol happened to be in the "ACL2" package
(where DEF-ERROR-CHECKER is defined), requiring uses of DEF-ERROR-CHECKER in
different packages to use ACL2::DESCRIPTION to refer to that formal (e.g. in the
TAILREC transformation, which is in the "APT" package). Now DEF-ERROR-CHECKER
uses as that formal the DESCRIPTION symbol that is in the same package as the
name of the function generated by DEF-ERROR-CHECKER.