Improve error-checking utilities.
Now these utilities use the recently added ER-SOFT-+ to potentially return more
informative error flags and values. They do so by taking the error flag and
value to return in case of error as extra arguments, between the DESCRIPTION
argument and the CTX argument.

The macro DEF-ERROR-CHECKER, via which most of the error-checking functions are
currently generated, has been modified to enable this extension. The manually
hand-written error-checking functions have been extended as well.

For now all the callers of the error-checking functions use T as the error flag
and NIL as the error value, so there should be no visible change in
behavior. However, the APT transformations will be extended to return more
specialized error information, to facilitate their programmatic
(vs. interactive) use.

Because ER-SOFT+ is in logic mode, some of the error-checking functions can
potentially be in logic mode. The DEF-ERROR-CHECKER macro will be extended with
an argument to choose between logic and program mode, and the error-checking
functions that can be in logic mode will be put in logic mode.

The error-checking tests have been adapted to the extension.