[x86isa] Sped up symbolic simulation by changing definition of program-at
Previously, program-at was defined in terms of combine-bytes.  It is
now defined in terms of a new function, combine-n-bytes.  This new
function is analogous to bitops::part-select, but it trafficks in
lists instead of numbers.  This change makes it straightforward to
prove lemmas like many-reads-with-rb-from-program-at from
rb-rb-subset.

Another benefit of this change is the speed-up in symbolic simulation
--- for instance, proofs/popcount/popcount-general used to take ~900s
to vertify, and now it takes ~300s.  This is because previously, the
RHS of many-reads-with-rb-from-program-at was absolutely grotesque:

(equal (mv-nth 1 (rb n addr :x x86))
       (logior (loghead 8
			(logtail (ash (+ addr (- prog-addr)) 3)
				 (combine-bytes bytes)))
	       (ash (mv-nth 1 (rb (+ -1 n) (+ 1 addr) :x x86))
		    8)))

This essentially caused many-reads-with-rb-from-program-at to be fired
again and again till we ran out of n. Yeah, grotesque.

The RHS of many-reads-with-rb-from-program-at is much nicer now:

(equal (mv-nth 1 (rb n addr :x x86))
       (combine-n-bytes (- addr prog-addr) n bytes))

During symbolic simulation of a program, we'd know the concrete value
of "bytes".  Even so, using combine-bytes instead of combine-n-bytes
would have been expensive because the former would combine all program
bytes whereas the latter only combines the necessary ones, i.e., "n"
bytes.