Prepare top-level X86ISA step function for 32-bit execution.
Now the execution of every instruction (except the unimplemented ones) splits on
whether we are in 64-bit or 32-bit mode. In the first case, execution goes as
before. In the second case, for now we stop with an 'unimplemented' message.

The next step is to turn the 32-bit mode unimplemented cases into something
implemented, by extending existing 64-bit semantic functions or in some cases by
providing new 32-bit semantic functions.

This commit breaks several proofs. The next commit fixes all these proofs.