Improve output of named formula utilities.
Each piece of information (when in verbose mode) starts with a blank line and is
followed by no blank line. (Before, it started with no blank line and ended with
a blank line.) This seems more in line with ACL2's general output; the
information being printed by these utility is text(-like).