Improve redundancy handling of APT transformations.
Now the :VERBOSE and :SHOW-ONLY option, which only affect screen output, do not
affect redundancy checks.

When a transformation is redundant and :SHOW-ONLY is T, the ENCAPSULATE
(retrieved from the transformation table) is shown on screen.

Instead of :REDUNDANT, a phrase is shown on screen if a transformaion is
redundant.