Improve TAILREC transformation.
Ensure that the target function is not the domain (if the domain is a function)
and is not called by the domain (if the domain is a lambda expression). This
seems reasonable since the purpose of the transformation is generally to replace
the old function with the new one, but this decision could be revisited if there
is a good example.