Change two backquotes to ordinary quotes, to work around apparent SBCL bug.
The bug was on linux, and was causing regressions to fail when books
attempted to include books/projects/x86isa/machine/x86.lisp.  It
didn't seem to happen with SBCL 1.3.11 but it did with SBCL 1.3.19 and
1.3.20.  I plan to send a bug report on Friday (as after considerable
effort, I've distilled the problem down so that ACL2 isn't necessary),
but this change suffices as a workaround.