Add utilities for paired names.
A paired name is a symbol consisting of (1) the first name of the pair, (2) a
global but customizable separator, and (3) the second name of the pair. Paired
names are useful, for instance, to denote theorems that state certain kinds of
relations between functions (denoted by the first and second names of the
pairs), where the separator describes the relation.