Restored rule standard-char-p-nth in axioms.lisp.
I accidentally deleted this rule, and an "everything" regression
passed without it.  But it's a nice rule so I've restored it.

Regressions performed without cleaning were failing without this rule
(as pointed out by Eric Smith), but that should be fixed now.

Ran only a "basic" regression so that I could get this in quickly.