Improve documentation of kwargs of defunc
First, in HTML, a <dl> tag's immediate children must all be either
<dt> or <dd>.  The <dl> describing the keyword arguments of defunc
contained a <p>, so I split it into two <dl>s and brought the <p>
outside.

Next, since <dd> tags are intended to be definitions of corresponding
<dt> tags, I reworked the section by moving "Keyword Options" out to a
<h4> before the <dl> and converting each <dd> into a <dt> + <dd> pair.

Finally, I reworded some of the text in the section, added a more
precise list of keyword arguments that are passed to the function
contract defthm form, and prettified some cross-package xdoc links.