Make defpointer support package-external symbols
Previously, defpointer took its arguments and ignored what package
they were in, incorrectly using the symbols with the same names in the
current package instead.  Now it respects the package in which its
arguments lie.

In making this change, I also had to decide what the link text of the
pointer should look like, depending on what package the linked doc
topic's symbol was in.  I decided that if the `from` and `to` symbols
were both in the same package, we should leave off the package
qualification in the text, and otherwise we should include it.

I also added a guard to the macro to ensure that only symbols are
accepted.