Fixed an ACL2(p) error that I apparently introduced in commit a680a466da987be381b0c93895f5cd336d5bcb2c.
In my editing of translate-in-theory-hint@par, I accidentally changed
eval-theory-expr@par to eval-theory-expr.  I've changed it back.
Note: since this is definitely an improvement, I did only the
equivalent of a "basic" regression before doing this commit/push.

Thanks to David Rager for pointing out an ACL2(p) build failure on
leeroy.