Add the APT transformation table utilities.
The table is used to record successful calls of transformations and to support
redundancy checks. The table may be extended with more information in the
future, and used for other purposes than just checking redundancy.

This is based on some code originally written by Eric Smith, which I have
extended and XDOCumented.