New book books/hints/use-pkg.lisp defines custom keyword hint for fleshing out :use hint.  New book books/system/acl2-system-exports.lisp defines *acl2-system-exports* to expand *acl2-exports*, and updated :doc accordingly.  Fixed a bug in redef-pkg book.
The utility in use-pkg.lisp is preliminary; email me if you want help
using it.  Thanks to David Rager for requesting it.

For *acl2-system-exports*, thanks to Alessandro Coglio for an initial
discussion leading to this change, and thanks to David Rager for
pointing me to the new "SMAN" package for some symbols to include.

Thanks to David Rager for pointing me to the redef-pkg bug.

NOTE: There's been some discussion about where to put the definition
of *acl2-system-exports*.  I put it in
books/system/acl2-system-exports.lisp, but others might want it under
books/std/package.lsp.  I figure it's easy enough to add the form
(include-book "system/acl2-system-exports" :dir :system) in
books/std/package.lsp; someone may want to do that.  This, way,
someone can include "system/acl2-system-exports" and not depend on a
books/std/ book.