`chmod a-x`, whitespace & indentation cleanup
I noticed there were some files in the repository that were marked as
executable despite not actually being executable programs or scripts.
This commit marks those as non-executable as they should be.

While doing this, I figured that if I was touching these files anyway,
I might as well do some whitespace and indentation cleanup on them as
well.  In some cases I expanded that whitespace and indentation
cleanup to other files in the same directory too.