Slightly improve form of the ABNF parser.
The parsing function for a rule of the form A = B / C was implemented as 'try
parsing C, otherwise try parsing B' (using SEQ-BACKTRACK). This is correct
because the ABNF alternation operator / is commutative, but now the parsing
function has been changed to try B first, similarly to the other parsing
functions. This is how a hypothetical parser generator would generate the
functions, so the new form is more representative of a possible generated
parser.

Some proofs had to adapted to this change, in particular "replacing" certain
lemmas with other ones.