Updating asdf to 3.3.0
asdf supports quicklisp.

To update asdf, I downloaded a new version of asdf.lisp from
https://common-lisp.net/project/asdf/ and replaced the version
of asdf.lisp in our repository with it.  I then rebuilt all the
books, after cleaning them.  This just checked that the new
version works on ccl.  We'll have to wait and see for feedback
from our ACL2 community to determine if there are problems with
other lisps.