Improve user interface of APT transformations.
The :VERBOSE input has been replaced by a :PRINT input that must be an APT print
specifier, which provides finer-grained control on the screen output produced by
the transformation.

Utilities and documentation for print specifiers have been put under
apt/utilities. The documentation of the individual transformations refers to the
PRINT-SPECIFIER topic for the :PRINT input.

The APT transformation table utilities have been updated to regard the :PRINT
input, instead of the :VERBOSE input, as irrelevant to redundancy (along with
the :SHOW-ONLY input).