Extend the X86ISA model of instruction fetching.
Extend instruction fetching to use the recently added function READ-*IP to read
the instruction pointer from RIP, EIP, or IP, based on mode (64-bit vs. 32-bit)
and (in 32-bit mode) CS.D (i.e. the D bit of the code segment).

So now instruction fetching is reading the correct instruction pointer not only
in 64-bit mode, but also in 32-bit mode.

All the proofs still worked, presumably thanks to the added theorem that reduces
READ-*IP to RIP in 64-bit mode.

It was also necessary to add, via DEFTHM-SB, theorems about the return type of
READ-*IP for a few proofs about instruction fetching to go through.

The next step is to use the recently added function INCREMENT-*IP to increment
the instruction pointer, making the increment correct not only in 64-bit mode,
but also in 32-bit mode.