Extend the X86ISA model with more 32-bit support.
Now the GET-PREFIXES function, which models the part of
instruction fetching that fetches instruction prefixes, uses the
newly introduced function INCREMENT-*IP instead of directly
incrementing the RIP by 1 (and ensuring that the result is a
canonical address). This way, now GET-PREFIXES operates correctly
in 32-bit mode too, besides 64-bit mode.

INCREMENT-*IP has also been slightly modified as follows:

- In 32-bit mode, when CS.D is 0 (i.e. the instruction pointer is
  16 bits), the incremented instruction pointer is no longer
  truncated to 16 bits. The reason is that, in well-formed states
  with well-formed segments, the limit of the CS segment will not
  exceed FFFFh, and so the segment limit check on the incremented
  instruction pointer will ensure that it consists of 16
  bits. Avoiding this truncation in the definition of
  INCREMENT-*IP keeps the function simpler.

- A more detailed error "flag" is returned if the incremented
  instruction pointer is not canonical (in 64-bit mode) or is not
  within the segment limit (in 32-bit mode). The additional
  details describes the error condition and the offending value,
  similarly to other functions in the model.

A few theorems and proofs had to be adapted to this change to
GET-PREFIXES (and to INCREMENT-*IP).