Re-organized a few files under misc.
The forms in misc/eval.lisp and misc/assert.lisp have been re-ordered so that
utilities are closed to their documentation (if any) and also so that related
utilities are close. Lines of semicolons ;;;;;;;... have been added to separate
different "sections" of these files.

The tests of ASSERT! and ASSERT!-STOBJ in misc/assert.lisp have been moved to
misc/assert-tests.lisp. Because of this move, misc/assert.lisp no longer needs
to include misc/eval.lisp, but since many files in the community books expect
misc/eval.lisp to be included automatically when misc/assert.lisp is included, I
am deferring to a later commit the removal from misc/assert.lisp of the
INCLUDE-BOOK of misc/eval.lisp.