Introduced :constrain argument for defun-sk, to make the function attachable and guard-verified.
Quoting :doc note-8-1:

  A new keyword argument for [defun-sk], :constrain, can specify that
  the newly-introduced function is constrained rather than defined.
  See [defun-sk].  Note that by constraining the function we make it
  possible to attach to it, and also to introduce it as a
  [guard]-verified function.  We also made a minor change to
  defun-sk, by moving the call of [extend-pe-table] out of the
  generated [encapsulate] form.

This feature should be useful for making warrants (for apply$)
executable and guard-verified, something we home to do soon.

NOTE: I've updated the utility defun-sk-check, to accommodate the move
of extend-pe-table out of the encapsulate generated for defun-sk.
However, it might be good to update those utilities so that they
comprehend the new :constrain keyword of defun-sk.  I'll discuss this
with Alessandro.