Add functions to manipulate the stack pointer.
The functions READ-*SP, ADD-TO-*SP, and WRITE-*SP serve to read and write the
stack pointer from and to RSP, ESP, or SP, and to increment or decrement it (by
passing a positive or negative delta to ADD-TO-*SP). These enable the correct
manipulation of the stack pointer not only in 64-bit mode, but also in 32-bit
mode.

These functions are not used by the rest of the model yet, but they will be used
by the semantic functions of instructions that manipulate the stack so that they
can be extended to 32-bit mode besides 64-bit mode.