Extend 'POP general-purpose register' to 32-bit mode.
The previous (64-bit-mode-only) version made a canonicality check on the stack
pointer read from RSP prior to reading the data in it. Now that 32-bit mode is
also allowed, the corresponding check would have been more complex (involving
segment limits). But since that check is made by RME-SIZE when reading the data
from the stack anyhow a few lines below, that check is no longer there.

Added a comment clarifying that the order of the updates to the stack pointer is
important when popping to the stack pointer (i.e. POP SP/ESP/RSP).