Extend POP r/m instruction towards 32-bit mode.
The writing to memory or register, carried out by X86-OPERAND-TO-REG/MEM, as
well as the address computation, carried out by X86-EFFECTIVE-ADDR, still need
to be extended to 32-bit mode.

The top-level dispatch is still considering POP r/m unimplemented in 32-bit
mode: this will be changed after extending X86-OPERAND-TO-REG/MEM and
X86-EFFECTIVE-ADDR to 32-bit mode.