VL: allow clocking events to be expressions
Previously, VL was unable to parse and represent the
SystemVerilog-2012 "sampled value system functions" (Section 20.13)
like $rose, $fell, etc., because they can take clocking_events like
``@(posedge clock)`` as an argument, and a clocking_event was not an
expression.  <rant>This is conveniently omitted from the alleged
grammar in Annex A, so it's no surprise that we missed it.</rant>

Fix this by extending the vl-expr representation to allow a new
"vl-eventexpr" case.  As a consequence, expressions become mutually
recursive with evatoms.  Extend the expression parse to handle this.
Add test cases to check that we can parse basic uses of these within
$rose calls.  Corresponding updates to the pretty-printer, sizing,
etc.

These expressions are not handled in any sensible way by, e.g., SV,
but should no longer result in parse errors.