Improve a function in X86ISA.
The function X86-EFFECTIVE-ADDR-FROM-SIB was supposed to return the displacement
as a separate result but instead it adds it to the rest of the effective address
and returns 0 as the displacement result. Following an email discussion with
Shilpi Goel, this function has been changed to return, not add, the
displacement.

The theorem asserting that the displacement result of this function is 0 has
been changed to show that the displacement result is an integer. The name of
this theorem has also been slightly changed to say INTEGERP instead of NATP,
because the displacement is signed. Also added to the documentation a reference
to where the AMD manual says that displacements are signed.