Extend X86-EFFECTIVE-ADDR towards 32-bit mode.
When Mod = 00b and R/M = 101b, the effective address is RIP-relative in 64-bit
mode, and displacement-only in 32-bit mode. The latter was previously
unimplemented, but it is now implemented. The final part of the code of this
function now also applies N32 to the final effective address in 32-bit mode, in
analogy with the pre-existing use of N32 in 64-bit mode when an address size
override prefix is present.

Now this function should cover 32-bit mode as well as 64-bit mode, if the
address size is not 16 bits. A separate function will be added to perform
effective address computation for 16-bit addresses (in 32-bit mode, because the
address size is never 16 bits in 64-bit mode), while the current
X86-EFFECTIVE-ADDR function will be used when the address is 64 or 32 bits (in
64-bit and in 32-bit mode). A "top-level" function to calculate the effective
address will dispatch to either function based on address size. This new
top-level function will be called X86-EFFECTIVE-ADDR, while the current one will
be renamed to reflect that it only applies to 64/32-bit addresses.