Generalize the SELECT-ADDRESS-SIZE function.
This was recently introduced. The generalization consists in having the function
take as argument a boolean flag indicating the presence of the address size
override prefix, instead of all the prefix bytes. The new version is more
general in the sense that it may be called from functions that may not have all
the prefix bytes. This change is so that this function can be called by
X86-EFFECTIVE-ADDR, which does not have all the prefix bytes.