Improve function to write stack pointer register.
Make the 16-bit writing consistent with explicit writing to general-purpose
registers -- the function that has been just improved is used by instructions
that implicitly write the stack pointer register (e.g. PUSH).

The Intel and AMD manuals could be clearer in this respect, but the consistency
mentioned above seems the most reasonable interpretation.