Fixed README.md title.  Added defpointers for PACKN and PACKN-POS.  Synched doc.lisp.  Fixed description for past supporting materials on workshops.html.
Changed "Community ACL2 System and Books" to "ACL2 System and Community Books"
in the top-level README.md.

Added defpointers for PACKN and PACKN-POS, which we try to do for all
symbols documented in system-utilities.

The workshops.html page had an out-of-date description for obtaining
past supporting materials for the workshops, as pointed out by Keshav
Kini, so I fixed that.