Expand ATOM in l3-to-acl2 translator's EQUAL-LEN-0
There are more than two dozen different versions of this theorem
throughout the community books.  Most of them rewrite (EQUAL (LEN X)
0) to (NOT (CONSP X)), some rewrite it to (ATOM X), and some rewrite
it to (ENDP X).  I'm not going to try to unify them all right now, but
@nadezhin wanted the one in the l3-to-acl2 translator to at least be
compatible with the one in the projects/apply/apply-lemmas book, which
happens to be the kind which rewrites to (NOT (CONSP X)).  Hence this
commit.