Add a library for number/digit conversions.
This library provides conversions between natural numbers and their
representations as digits in arbitrary bases in big-endian and little-endian
order. The digits are natural numbers below the bases.

The conversions have a tail-recursive definition for execution and a
natural-recursive definition for logical reasoning.

The library includes, among others, theorems stating that number-to-digits and
digits-to-number conversions are mutual inverses (modulo fixing, and modulo the
most significant zeros, as appropriate).

Also add a small library for lists of ZPs, used by but independent from the
number/digit conversion library.