Extend other ADD/ADC/etc. variants to 32-bit mode.
These are the (Eb Ib), (Ev Iv), and (Ev Ib) variants of the instructions ADD,
ADC, SUB, SBB, OR, AND, XOR, and CMP.

The same semantic function that has just been extended also implements the Eb
and Ev variants of the TEST instruction, which will be enabled at the top level
soon.

Prior to these extensions, the semantic function was generating #UD when the
LOCK prefix is used and the instruction is CMP. I relaxed the test for the CMP
instruction to include the TEST instruction as well, because that one also
generates #UD with the LOCK prefix. The new function also generates #UD when the
LOCK prefix is used but the destination operand is a register and not memory.