Extend SELECT-OPERAND-SIZE to 32-bit mode.
This function now takes the x86 state as an additional parameter, from which it
extracts the mode (64-bit or 32-bit) and the CS.D bit. These are used, together
with the other parameters, to determine operand size in both 64-bit and 32-bit
modes.

Since the XDOC page of SELECT-OPERAND-SIZE contained an image capture of Table
3-4 of Intel manual Volume 1 for operand (and address) size determination in
64-bit mode, I also added an image capture of Table 3-3 of Intel manual Volume 1
for operand (and address) size determination in 32-bit mode.