Added argument :last-body to make-flag.  Improved install-not-normalized to handle non-recursive normalized bodies of recursively defined functions, and fixed copy-def in that case.
Quoting :doc note-8-1-books:

  The utility [make-flag] has a new keyword argument, :last-body, to
  specify using the most recent [definition] rule for a function
  symbol instead of its original definition.

  Improved [install-not-normalized] to handle cases in which
  recursively-defined functions have non-recursive normalized
  definitions.

Also, improved copy-def to work better with mutually recursive
definitions with non-recursive normalized bodies.  Added a test in a
new book, books/kestrel/utilities/copy-def-tests.lisp, that failed
before this improvement.  The copy-def improvement takes advantage of
the imrpvoement to make-flag by calling it with :last-body t, so that
it uses definition rules just installed by install-not-normalized, and
it also takes advantage of the improvement to install-not-normalized.

In support of the change to install-not-normalized.lisp, the
all-fnnames1 is now a guard-verified :logic mode function.  (Note that
calls of the macros all-fnnames and all-fnnames-lst expand to calls of
all-fnnames1.)