Fixed download by adding index.html to the download/ directory.
Changes to fix download button:

- Created new file books/xdoc/fancy/download-index.html.

- Modified books/xdoc/fancy/zip.sh so that it copies
  books/xdoc/fancy/download-index.html to
  books/doc/manual/download/index.html.

- The download button now links to download/index.html instead of
  download/.

Thanks to David Rager for suggesting the addition of
download/index.html as a way to get the download button to work, and
to Jared Davis for seconding that idea.

The new download/index.html, i.e.,
books/xdoc/fancy/download-index.html, is very simple.  Anyone who
thinks it's ugly is welcome to improve it!