Improve APT transformations.
Check that only hints for applicability conditions that are present (based on
other inputs to the transformation) are supplied in the :HINTS input. For
instance, if the generated function(s) are not to be guard-verified, the
applicability conditions related to guard verification are absent and thus the
coresponding hints must be absent too.