Improve XDOC constructors.
Open the door for changing the implementation to something else than the current
one based on string concatenation. So far string concatenation has not been too
slow in the uses of these XDOC constructors, but if it turns to be too slow in
the future, it should be now possible to change the implementation with relative
ease, without disrupting the code that uses these XDOC constructors.