Warn for definition rules with non-nil :install-body for functions in *definition-minimal-theory*.
A recent commit removed the restriction that :install-body must be nil
for functions in *definition-minimal-theory*.  The change was to avoid
causing an error when that restriction failed.  Now, we issue a
warning where there was originally an error, pointing out that calls
of any such function will often be expanded using the original
definition of the function symbol.