Extend INC and DEC to 32-bit mode.
These are the variants of INC and DEC with opcodes FEh and FFh.

The variants with opcodes 40h-4F, which are only valid in 32-bit mode, were
already added to the model.

We may consider merging the models of these two variants of INC/DEC, but for now
there are separate semantic functions for them.