Extend DEFBYTE with more keyword arguments.
The new keyword arguments :LTYPE, :LPRED, :LFIX, and :LEQUIV are analogous to
the ones without the starting 'L': these apply to the fixtype, recognizer,
fixer, and equivalence for lists of bytes (as opposed to the ones for bytes).

The default recognizer name adds -P for consistency with other FTY and STD
macros, so I have added explicit :LPREDs to generate ...-LISTP recognizers
instead for the existing calls of DEFBYTE, following the CLtL2 convention that,
for lists, is also followed in the ACL2 sources and libraries.