Abbreviate computed hint message
The function defined by automate-instantiation computes a hint, prints
it as a "suggestion" to the user, and then immediately provides that
same hint to the theorem prover.

Since the hint is being provided to the prover anyway, there doesn't
seem to be a need to also tell the user about it, and these hints are
often very large and contribute a lot of noise to the proof output
when trying to prove a large theorem that uses pick-a-point subset
reasoning a lot.

So, this commit greatly abbreviates the message that the computed hint
prints.  The new message is only one or two lines long, and still
tells the user how to disable the hint (in case the pick a point
strategy is ruining the proof), but doesn't supply any further
information.